backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
opencv-python==4.11.0.86
numpy==2.2.4
scikit-learn==1.6.1
Pillow==11.2.1
